semana2-02

+++ src/control/AlmacenArticulo.java
@@ -44,15 +44,20 @@ public class AlmacenArticulo<T> {
 		if (Utiles.comprobarExiste(pathIndice)) {
 			indice = (TreeMap<String, Integer>) new DAO().leer(pathIndice);
 			Integer nombreArch = indice.get(nombre);
-			pathDatos.append(nombreArch + ".art");
-			retorno = (T) new DAO<>().leer(pathDatos.toString());
-			this.pathDatos.delete(lenghtDir, pathDatos.length());
+			if (nombreArch!=null) {
+				pathDatos.append(nombreArch + ".art");
+				retorno = (T) new DAO<>().leer(pathDatos.toString());
+				this.pathDatos.delete(lenghtDir, pathDatos.length());	
+			}
+			
 		}
 		return retorno;
 	}
 
 	public TreeMap obtenerIndice() {
-		return (TreeMap) new DAO<>().leer(pathIndice);
-		
+		if (Utiles.comprobarExiste(pathIndice)) {
+			return (TreeMap) new DAO<>().leer(pathIndice);
+		}
+		return null;
 	}
 }
